+++ base/common/src/com/netscape/cms/servlet/cert/DoUnrevokeTPS.java
@@ -549,7 +549,7 @@ public class DoUnrevokeTPS extends CMSServlet {
 
         BigInteger[] biNumbers = new BigInteger[biList.size()];
         for (int i = 0; i < biList.size(); i++) {
-            biNumbers[i] = (BigInteger) biList.elementAt(i);
+            biNumbers[i] = biList.elementAt(i);
         }
 
         return biNumbers;
